Show selection options in details component
On the readonly questions page for a live form, show the options for
a "Select from a list of options" question in a details component if
there are more than 10 options.

The count for the number of options includes the "None of the above"
option, if one is present, as we include that in the list.