Exclude /up from HostAuthorization check
The HostAuthorization middleware is used to protect against DNS
rebinding attacks.

We want to set config.hosts to validate the Host header of incoming
requests. This is a security feature that prevents an attacker from
making a request to a server with a different hostname than the one
that the server is actually running on.

The HostAuthorization middleware is used to protect against DNS
rebinding attacks. It checks the Host header of incoming requests
against a list of allowed hostnames. If the hostname is not found in
the list, the request is rejected.

Doing this will make the health check used during our deployments to fail
because AWS will use an internal IP address instead of the public DNS
entry, which is still running the previous version of the app.

Here we disable the HostAuthorization middleware for the health check
route. This means this route is still vulnerable to DNS rebinding. I
don;t think that matters for us thought. The route only ever returns a
status code and never redirects.

There are other options. For example, adding internal IP ranges to the
allow list. This complicates configuration though and leaves us
potentially open in other ways.

I think this is probably our best option.

https://guides.rubyonrails.org/configuring.html#config-hosts
https://api.rubyonrails.org/classes/ActionDispatch/HostAuthorization.html