Use bulleted list for selection options
As we do on the edit question page, display the list of options for a
selection question as a bulleted list.

When there are 10 or fewer options and we do not use the details
component, have a paragraph before the list stating the number of
options as we do on the edit question page.