Add default_group_id to ignored_columns
To prevent any cache issues, we need to add default_group_id
to the ignored_columns list before dropping the column.

The column is no longer used throughout the app. After this commit any
use will raise an error.

The column will be removed in a later PR.

See this these links for more information:

https://api.rubyonrails.org/classes/ActiveRecord/ModelSchema/ClassMethods.html#method-i-ignored_columns-3D
https://github.com/ankane/strong_migrations?tab=readme-ov-file#removing-a-column