Make sure FormRepository.pages is stubbed in tests
In a number of tests we have not been mocking FormRepository.pages, but
it has been working anyway because the form record has `pages` in its
attributes (so ActiveResource does not make a network request).

However, this makes it harder to do our migration to API v2, because we
want our specs to work regardless of how FormRepository is implemented.

In a previous commit we added code to raise an exception if
FormRepository.pages was being called in a test with a form record that
had a `pages` attribute.

This commit stubs FormRepository.pages everywhere this
exception is being raised in our tests.