Add validation to privacy URL to prevent example
When form creators are asked to enter a privacy policy URL, they are
shown an example of a privacy policy URL. This URL is not suitable for
forms to use, so we should prevent them from entering it.

This commit adds a validation to to prevent the example URL from being used.

Co-authored-by: Hannah Content <hannah.cooper@digital.cabinet-office.gov.uk>