Ensure external ID is present in database for forms
In forms-api forms have an internal ID which is the database primary
key, and an external ID which is called ID in the REST API. For
migrating form objects from forms-api to forms-admin we have been using
the response ID as the primary key, as at the moment the primary key and
external ID have the same value in all cases. However we had forgotten to
save the external ID to the forms-admin database.

This commit adds the external ID to the `#database_attributes` method
for form resources.