Change Api::V1::PageResource#move_page to update page in place
Use the protected method
`ActiveResource::Base#load_attributes_from_response` to update the page
when the `#move_page` method is used; this means that after calling
`#move_page` the page object will have the updated position attribute.

This fixes a bug we were seeing in production with the PageRepository
class when trying to use `from_json` to load the attributes, as
`load_attributes_from_response` works better (see previous commit).