Make active record model factories similar to active resource factories
We want to start returning Active Record models from our forms models
repositories. This means we also need to swap out the factories used in
our specs, and to allow that we need to make sure that the same traits
are present with the same behaviours across both sets of factories.

This commit makes the `*_record` factories more similar to the
`*_resource` factories, by adding missing traits and tweaking the
existing ones where there is a mismatch. In some cases we can't use the
exact same code to implement the trait, and for those we have added
tests of the traits.