Add DataStruct, DataStruct.recursive_new, and DataStructType
Add a DataStruct type to treat JSON columns in a record as objects that
have attributes that can be accessed using dot `.` instead of hash
access `[]`.

We need this because ActiveResource turns nested objects into classes
with attributes, and our existing code assumes that attributes can be
accessed by writing `.attribute_name`.

Co-authored-by: Stephen Daly <stephen.daly@digital.cabinet-office.gov.uk>
Co-authored-by: Samuel Culley <samuel.culley@digital.cabinet-office.gov.uk>
Co-authored-by: Laurence de Bruxelles <laurence.debruxelles@dsit.gov.uk>