Explicitly configure vite cache directory
Something (possibly a recent Vite update) made vitest read the cache
directory as relative to the frontend folder.

This meant that it was generating unwanted cache files in
`/app/frontend/node_modules/.vite`.

This commit explicitly configures the cache folder to be the expected
`/node_modules/.vite`, which is gitignored, to avoid this.