Fix tests passing with false positives
The arguments to a Rails path helper should be either an object such as
a record that implements ActiveModel::Naming ActiveModel::Conversions,
or a keyword argument [[1]]; if an object `o` is passed as an argument
then `o.model_name.route_key` will be used to work out what parameter in
the path `o.to_param` should be used for, and if a keyword argument is
given then the name of that keyword will be used for the path parameter.

If an object is given as a keyword argument, the model name is not used
for the path parameter; only the keyword argument name is used.

These tests have a subtle bug, where the record objects were mistakenly
passed as keyword arguments to path helpers in the spec expectations.
The tests passed because the value of `#to_param` was `nil`, because the
records were not persisted (see the definition of
ActiveModel::Conversion#to_param [[2]]), and thus were ignored by the path
helper, but the correct path parameters were accessible from the
request. If the records had been persisted then a URL parameter named
after the keyword would have been appended as a query parameter; not the
desired behaviour.

This commit changes these tests so that they all use keyword arguments
for path helpers; they are more explicit and allow the tests to pass
for the right reason.

Note that there is one test failing after this change; it is failing
correctly now, previously the bug was causing it to pass incorrectly.

[1]: https://guides.rubyonrails.org/routing.html#creating-paths-and-urls-from-objects
[2]: https://api.rubyonrails.org/classes/ActiveModel/Conversion.html#method-i-to_param