Fix path in redirect after deleting page
The path helper for the redirect had the ID of the deleted page passed
to it, but did not have a slot for a page ID in the path pattern, so the
URL ended up with a random number appended. There was a test, but it was
passing incorrectly (see the previous commit), so this was previously
missed. It's fixed now.