Don't set next_page in specs
Rather than setting `next_page` on page objects, create a form with a
list of pages and rely on the behaviour of the form factory/acts as list
gem to link the pages together and set `next_page`.

This is needed because when using ActiveRecord models, `next_page` is a
method rather than an attribute, and it is computed from the page
position.