Fix conditions_controller_spec#create
The test was unnecessarily invoking the factory method to create a
condition when stubbing the ConditionsResource#create method. This
broke as the factory is now using the ActiveRecord factory. We didn't
need to invoke this - instead just return any condition.

Add an additional spec to assert that the ConditionsResource#create
method is called.

Co-authored-by: Samuel Culley <samuel.culley@digital.cabinet-office.gov.uk>
Co-authored-by: Laurence de Bruxelles <laurence.debruxelles@dsit.gov.uk>