Add payment links to live view of form
When displaying a live form, we don't show any information about payment
links.

This commit adds a payment link to the live view of a form if one is
present.

We have to be slightly careful to not rely on payment_url? as it may not
exist for forms published before we added the payment_url field.