Enable using Spring to speed up RSpec startup
Uses [Spring] with the [RSpec command] to reduce the amount of time
RSpec needs to load files when running tests.

To use Spring run `bundle exec spring rspec`. This will start a server
in the background, speeding up subsequent runs of the command. The
server will automatically when the terminal is closed, but it can also
be stopped manually with `bundle exec spring stop`. See the [Spring]
readme for more details.

The workaround for segfaults is from [[1]].

[Spring]: https://github.com/rails/spring
[RSpec command]: https://github.com/jonleighton/spring-commands-rspec

[1]: https://github.com/ged/ruby-pg/issues/311#issuecomment-561927000