Disable test coverage measuring by default to speed up RSpec
Make coverage with [SimpleCov] optional unless the `COVERAGE` environment
variable is set. Now the coverage information in `./coverage` will not
be updated unless tests are run as

`COVERAGE=true bundle exec rspec`

This helps with reducing the amount of time for tests to load. By itself
the coverage code doesn't slow down the tests that much, however the
interaction with [Bootsnap] caching impacts load times, because Bootsnap
does not cache modules if the Coverage code is in use [[1]].

[Bootsnap]: https://github.com/Shopify/bootsnap
[SimpleCov]: https://github.com/simplecov-ruby/simplecov

[1]: https://github.com/Shopify/bootsnap/issues/35