Fix QuestionsController to handle Page ActiveRecord model
The `page` loaded by the QuestionsController is now a page
ActiveRecord model instead of an ActiveResource model. This means that
we need to update attribues on the model using the `assign_attributes`
method instead of the `ActiveResource::Base#load` method.

Updates the tests in questions_controller_spec to work with the
changes on this branch to use ActiveRecord models instead of
ActiveResource models at the same time.

Co-authored-by: Stephen Daly <stephen.daly@digital.cabinet-office.gov.uk>
Co-authored-by: Samuel Culley <samuel.culley@digital.cabinet-office.gov.uk>