Fix ConditionsInput to work with Page ActiveRecord model
Previous commits on this branch mean that the code now deals with Page
ActiveRecord models instead of Page ActiveResource models. This commit
updates the `ConditionsInput` class to work with the Page ActiveRecord
model, where the `answer_settings` is a DataStruct rather than an
ActiveResource model with an `attributes` property.

This change also updates the `conditions_input_spec` to ensure that it
works correctly with the new ActiveRecord models instead of
ActiveResource models.

Co-authored-by: Samuel Culley <samuel.culley@digital.cabinet-office.gov.uk>
Co-authored-by: Stephen Daly <stephen.daly@digital.cabinet-office.gov.uk>