Change ConditionRepository to accept and return ActiveRecord Conditions
The condition repository now takes in ActiveRecord objects instead of
ActiveResource objects. Find and Create methods now return
ActiveRecord objects, which now means that the save! and destroy!
methods expect to receive ActiveRecord objects.

The API is still involved, as all repository methods will first send a
request to the API, and use the result of that to update the db.

Co-authored-by: Samuel Culley
<samuel.culley@digital.cabinet-office.gov.uk> Co-authored-by: Stephen
Daly <stephen.daly@digital.cabinet-office.gov.uk>