Handle nested ActiveResource::Base when converting PageResource
When we convert a Api::V1::PageResource to a Page ActiveRecord model,
we cast the `answer_setttings` attribute to a DataStruct. This cast
was not working because the `answer_setttings` has type
Api::V1::PageResource::AnswerSettings which is a subclass of
ActiveRecord::Base.

Update the cast method for DataStructType to handle ActiveRecord::Base
to allow for this conversion.

Co-authored-by: Stephen Daly <stephen.daly@digital.cabinet-office.gov.uk>
Co-authored-by: Samuel Culley <samuel.culley@digital.cabinet-office.gov.uk>