Fix the pages list error summary component to work with ActiveRecord
Fix the error summary component for the pages list to work with
ActiveRecord Page model instead of the ActiveResource models.
Previously, we had to iterate over the Conditions returned by
ActiveResource to resolve the check page per condition to use in the
error message. Now that the conditions are loaded by ActiveRecord,
they already have the check page assigned so this is unnecessary.

We've fixed the specs for the component at the same time to work with
ActiveRecord models instead of ActiveResource models.