Fix preview_components_spec
preview_components_spec attempts to load each component preview and
then do a check that there are no accessibility errors.

These started failing for the page_list_component_preview and the
page_list_component_preview while we are switching over to using
ActiveRecord instead of ActiveResource form the Form/Page/Condition
models.

The previews actually worked fine when the server is run and they are
viewed in the browser, but not when they were loaded by the
preview_components_spec. This is because the associations were not set
for the models created using the factories.

We cannot switch the previews to use `create` rather than `build` as
we do not want to persist data to the database when loading the
previews in the browser. Also, when attempting to use `create`, these
tests don't seem to be able to select IDs that are not already in use
automatically.

The solution is to set the associations manually on the Conditions
created, so they can be accessed as needed to load the components.