Fix DataStruct conversion to deal with nested ActiveResource object
The answer_settings for the PageResource ActiveResource model is an
object that inherits ActiveResource::Base, which itself can contain an
array of ActiveResource::Base models if it has selection options.

Fix our code to cast the anwswer_settings into DataStruct objects to
handle the nested ActiveResource::Base models for the selection
options.

Co-authored-by: Samuel Culley <samuel.culley@digital.cabinet-office.gov.uk>
Co-authored-by: Laurence de Bruxelles <laurence.debruxelles@dsit.gov.uk>