Fix secondary_skip_controller_spec to work with ActiveRecord models
In previous commits on this branch we:

* Updated the FormRepository, PageRepository and ConditionRepository
to accept and return ActiveRecord models instead of ActiveResource
models.
* Updated the form, page and condition factories to build ActiveRecord
models instead of ActiveResource models.

These changes caused most specs that deal with forms/pages/conditions
to start failing. This commit makes fixes to the
secondary_skip_controller_spec to work with the new ActiveRecord
models.

Co-authored-by: Laurence de Bruxelles <laurence.debruxelles@dsit.gov.uk>
Co-authored-by: Samuel Culley <samuel.culley@digital.cabinet-office.gov.uk>
Co-authored-by: Stephen Daly <stephen.daly@digital.cabinet-office.gov.uk>