Fix creating draft question from existing page
When creating a draft question from an existing page, we want to copy
across all of the existing page's attributes except for its id. As part
of the change over to using ActiveRecord objects, the way we identify
the id using a symbol meant that we weren't correctly excepting it from
our draft question attribute assignment.

This has been changed into a string, so that it is correctly excepted
when creating a draft question.

Co-authored-by: Samuel Culley <samuel.culley@digital.cabinet-office.gov.uk>
Co-authored-by: Stephen Daly <stephen.daly@digital.cabinet-office.gov.uk>