Fix wrong page number issue when deleting routes
We were seeing the wrong page number appearing when the routes for a
page were deleted.

This was happening because our page_number method was being too strict
when looking for the page in the form. If a page had changed any of its
attributes, the method would conclude that the page did not exist, and
would return a new page number (the length of the form, plus 1).

This was happening on the delete routes page because in the #destroy
action:

- the page object is memoized
- the routes are deleted from the page
- the success message then calls the page_number method with the
  memoized page object
- the page_number method compares the memoized page object with its
  current state

This commit resolves the issue by updating the `page_number` method to
only compare the IDs of pages. This should make it more resilient to
being used with memoized pages like this case, and should still work
since page IDs are unique.