Make nil check explicit
If page.id is nil, we'd expect the page_number method to return a 'new'
page number (i.e. pages.length + 1).

This already happens because there shouldn't be any page with nil id in
the form, so the comparison in the pages.index should return nil.

This commit makes that behviour more explicit with a guard clause, and
also adds a test case for this behaviour.