Add move_page method to Page model
This method replicates the behaviour currently in forms-api to move a
page up or down in a form.

This differs slightly in implementation because it will always update
the form's state and create a draft if the form is live or archived
even if the page hasn't moved because it was the first or last page
already. We won't show buttons to move the page if it is the first or
last page, so this will only result from a user re-submitting a
request and we don't think it matters that we'll update the form state
in this case.

Co-authored-by: Samuel Culley <samuel.culley@digital.cabinet-office.gov.uk>
Co-authored-by: Laurence de Bruxelles <laurence.debruxelles@dsit.gov.uk>