Change behaviour of FormRepository with use_database_as_truth
For the find method, when use_database_as_truth is true, we want to
just retrieve the form from the database and not make any request to
the API.

For all other methods that create/update a form, we want to first
update it in the database and then send the updated record to the API.

The where method will only use the local database to source its data

The make_live! method will not modify the local form's page data

The archive! method does not change behaviour with the flag

The destroy method doesn't need to change based on the setting,
however we improved it to ensure we always attempt to destroy from the
database even if the form has already been deleted from the API.

The pages method does not modify local page data or call the API

The behaviour is unchanged when use_database_as_truth is false.

Co-authored-by: Stephen Daly <stephen.daly@digital.cabinet-office.gov.uk>
Co-authored-by: Laurence de Bruxelles <laurence.debruxelles@dsit.gov.uk>