Set the next value in the sequence for primary keys for forms tables
Because we have been upserting with IDs into the forms, pages and
conditions tables, the primary key sequence is still at 1 for these
tables. This means that when we try and create a record in one of these
tables without specifying an ID, we are likely to get a duplicate ID
error.

This commit sets the sequence to start at a value higher than the
current highest ID in each table.

To make it less likely that the IDs for records created in the API
database to clash with IDs for the admin database, a suitably large
number is chosen leaving a gap for safety margin.