Change primary key sequence for forms tables to be less weird looking
We found after some testing that having IDs that started at a high power
of 10 looked weird because they had a lot of 0s; we worried that this
would result in a negative user experience.

Instead set the primary key sequence to start from a high base 2 number,
this has the advantage of having a more even distribution of digits
(looking less weird and having fewer 0s), whilst still being
recognisably different that we can see at a glance that the change has
worked when needed.