Fix move page flash message position
When a page is moved, we show a success message to indicate the new
position of the page. As part of change to the way the Page Repository
moved pages around, we were unecessarily recalculating the new page
position in the controller.

This change makes use of the new Page Repository move_page behaviour to
get the new page position directly from the updated db page.

Co-authored-by: Laurence de Bruxelles <laurence.debruxelles@dsit.gov.uk>