Display correct content for date of birth answer type
Previously, on the live questions page, we were displaying the answer
type as "Yes" for date questions where “Are you asking for someone’s
date of birth?” was answered as true when creating the question.

Fix this so that we display the answer type as "Date of birth".
Continue displaying the answer type as "Date" for date questions that
aren't asking for a date of birth.