Add #as_form_document to form
Returns a json blob of a form, which includes the form's pages as steps.
It also removes a few attributes which aren't necessary for Forms
Runner, mainly to do with checking the state of a form, and its completion
tasks.

The JSON that it outputs matches the current Forms API output for a form
document, to prepare for making form documents in Admin instead.

Co-authored-by: Stephen Daly <stephen.daly@digital.cabinet-office.gov.uk>