Prepend empty line to markdown block elements
Markdown block elements (in our case, headings and list items) should
have an empty line between them and any previous content.

If this empty line isn't present, it's possible the markdown will not be
rendered as the user intended.

This commit adds logic so that if the user adds a block element using
the markdown toolbar, they'll automatically have an empty line added if
there isn't already one present.