Add association for getting forms for a group
Link Forms to Groups through associations on the Group, GroupForm and
Forms models. Use this association to retrieve forms for a group
instead of using the FormRepository.

We no longer destroy the GroupForm when the FormResource is destroyed,
and instead added `dependent: :destroy` to the new association on the
Form so that the GroupForm is destroyed when the Form is destroyed.

This should be quicker!

Co-authored-by: Samuel Culley <samuel.culley@digital.cabinet-office.gov.uk>
Co-authored-by: Laurence de Bruxelles <laurence.debruxelles@dsit.gov.uk>