Create/update form document when form changes state
When a form goes live or gets archived, we want to create a FormDocument
that represents the form at that point. When going live, we use the
current state of the form in the db to determine the content for the
form document.

When archiving a form, we want to use the existing live form document's
content as the archived form doc's content. We may not yet have a live
form document in the database, so if we don't we pull it from the API,
and create an archived form doc using the content from there.

Co-authored-by: Laurence de Bruxelles <laurence.debruxelles@dsit.gov.uk>