Use group form association to get forms for show group action
Now that we have an association through group_forms connecting groups
and forms, we can use the database to collate the forms for a group
rather than looping through each group_form in turn.

This should make the view to list the forms for a group quicker for
groups with lots of forms.

This was missed from PR #2159.