Add EMAIL environment variable to review app
When seeding the database, the seed needs to draw an email from either
the EMAIL environment variable, or from the user's git login - which
isn't available in review apps. We need to set an EMAIL env var when
setting up a review app so that it can populate the db accordingly.

Co-authored-by: Laurence de Bruxelles <laurence.debruxelles@dsit.gov.uk>