Rename live controller local to form_document
The live and archived controllers both pass a local which was named
"form", which is slightly misleading as it is not passing through a Form
object - which means that it does not behave like a Form object in
certain circumstances.

We will be making a change to use local FormDocument data instead of API
data, so we are making this change in anticipation of this.

Co-authored-by: Stephen Daly <stephen.daly@digital.cabinet-office.gov.uk>
Co-authored-by: Laurence de Bruxelles <laurence.debruxelles@dsit.gov.uk>