Add type="module" to analytics script
We were getting a 'Config command out of order' warning in Google Tag
Manager because our analytics script was sending events before the GTM
config script ran.

This may result in unexpected behaviour while routing and processing
those events and might be a cause of the behaviour we're seeing with
validation errors not being properly populated in Google Analytics.

This was happening because the main application.js file, where the GTM
tag is configured, has `type="module"` but the analytics script is a
plain inline script. The execution of module scripts is deferred until
after the HTML is parsed, but inline scripts are executed immediately.

This commit adds `type="module"` to the inline script so that it will be
deferred until after the HTML has been parsed and the application.js
file has executed.