Remove repeated group and organisation lookups
The reports were joining on the group and organisation for the form
when doing the initial query to get the form data in order to filter
out forms from internal organisations. They were then later looking up
the group and organisation per form in order to construct the data for
either the table of forms/questions to display or the CSV to download.
Remove the lookup of group/organisation per form by selecting the
group and organisation details in the initial query and using this
when rendering.

Some tests were removed from report_helpers_spec for when the form is
not in an organisation, but we no longer include forms without an
organisation in the reports, so these were unnecessary.

Co-authored-by: Samuel Culley <samuel.culley@digital.cabinet-office.gov.uk>