Refactor logging specs
We often want to capture the output of the Rails logger to check whether
some information has been logged; in a fair few of our request specs we
mock the logger in the same way, resulting in duplicate code.

This commit makes it easier to add logging examples to a spec file,
instead of copying and pasting several lines we can now just add the
`:capture_logging` tag to the relevant context block and then the
`log_lines` and `log_line` helpers will be present in the examples.