Update FormDocumentSyncService to not call API
Removes the step in the form document sync that involved checking the
API for an existing live form if one didn't exist locally.

This is no longer needed as we have run a task to copy all live
FormDocuments from forms-api into the forms-admin database.

Instead, raise an error if there isn't a local "live" tagged form
document.

Co-authored-by: Samuel Culley <samuel.culley@digital.cabinet-office.gov.uk>