Update FormDocumentSyncService to handle multiple FormDocuments
We plan to start saving draft FormDocuments, and so there will
now be multiple FormDocuments per Form. This means we can no longer
get any existing FormDocument to update when a form is made live or
archived, and instead we must get the existing FormDocument with the
same tag (live or archived) to update.

We also need to make sure we delete live FormDocuments when a form is
archived and delete archived FormDocuments when a form is made live.

Co-authored-by: Samuel Culley <samuel.culley@digital.cabinet-office.gov.uk>