Add rake task to create draft FormDocuments
Draft FormDocuments will now be created each time a Form is updated.
We need to backfill existing Forms with draft FormDocuments so that
we can use these when serving the API response to get the draft
FormDocument and for the draft form feature reports.

Add a rake task to create/update draft FormDocuments for all existing
Forms. We have decided that draft FormDocuments will exist for all
Forms even if they are in "live" or "archived" status without any
changes to their draft.