Use draft FormDocuments for reports
We draft FormDocuments in the database and update them whenever a Form
is updated. Update the feature reports to retreive draft FormDocuments
from the database instead of generating them on the fly. The query to
get FormDocuments for draft forms will only include Forms that have a
`state` indicating that there are draft changes, so will not include
forms in state `live` or `archived`. This is consistent with the
existing behaviour.