Read FormDocuments for reports in batches
To avoid loading all the FormDocuments into memory at once, we now
read them in batches of 100. This probably isn't an issue now, but may
become an issue as we get more forms.

To acheive this, use `find_each` in FormDocumentService#form_documents
to return an Enumerator. Also call the `Enumerator#lazy` method so we
don't enumerate the data until we need to.

The batch size of 100 was picked somewhat arbitrarily. It could be
adjusted if needed.