Reduce the number of queries for the groups index page
Previously, an SQL query was executed for each group to get the user
that created the group.

Instead, we now retrieve all the creators in a single query and have
a method to pull out the creator by the ID. This is the same strategy
we use on the group forms page.