Handle "archived_with_draft" state in FormDocumentSyncService
We weren't handling the case where the form status is updated to
the "archived_with_draft" state, which is the state we transition to
when archiving a form in the "live_with_draft" state. This should be
handled the same way as the "archived" state, creating a FormDocument
with the "archived" tag and deleting the live FormDocument.