Fix translations for summary list on delete condition page
A block passed to
`GovukComponent::SummaryListComponent::RowComponent#with_key` is
evaluated in the context of that class, rather than the context of the
view; so for a translation key starting with `.` the scope will
be `govuk_component.summary_list_component` rather than the scope for
the view as expected.

Testing locally it seems the behaviour changed in govuk-components
v5.11.2; this was the release that bumped view_component from v2 to v4,
so I think the breakage was probably introduced by that gem.

It's not clear whether the new behaviour is intended or
not, but we can workaround easily enough by passing the desired string
as an argument to `with_key` rather than using a block.