Render 404 page when form is in the wrong state
When GET requests are made to /forms/:id/live and /forms/:id/archived
routes, check that the form is in the correct state (live or archived)
to render the page. If not, render a 404 page.

Previously, no check was made and an unhandled error was raised when
trying to load the form due to a FormDocument with the desired tag not
existing.