Add separate ErrorsController for API error handling
Forms admin now serves API requests as well as web requests. If
application/json is requested for a request, errors should be returned
with content type application/json.

Add a separate ErrorsController to render errors as JSON. Route to
the Api::ErrorsController if the request format is JSON, otherwise
route to the existing ErrorsController.

Set config.consider_all_requests_local = false in test.rb so that
Exceptions are not intercepted and rendered with a stack trace by the
Rails exception handling middleware in tests.