Add an 'All organisations' option to the user filter
Add an option to the "Organisation" filter on the user index page that
allows clearing the filter and showing users from all organisations.

It isn't possible to insert an option with a blank value into the
autocomplete, so instead set the value to 'all' and handle that case
in user scope.